updated plugins and removed default themes

+++ wp-content/plugins/wp-hummingbird/core/object-cache.php
@@ -2,16 +2,16 @@
 /**
  * Hummingbird Redis Object Cache
  *
- * @link    https://premium.wpmudev.org/project/wp-hummingbird/
+ * @link    https://wpmudev.com/project/wp-hummingbird/
  * @since   2.5.0
  * @package Hummingbird
  *
  * @wordpress-plugin
  * Plugin Name:       Hummingbird Redis Object Cache
- * Plugin URI:        https://premium.wpmudev.org/project/wp-hummingbird/
+ * Plugin URI:        https://wpmudev.com/project/wp-hummingbird/
  * Description:       Hummingbird object cache powered by Redis.
  * Author:            WPMU DEV
- * Author URI:        https://premium.wpmudev.org/
+ * Author URI:        https://wpmudev.com/
  *
  * Based on Eric Mann's and Erick Hitter's Redis Object Cache: https://github.com/ericmann/Redis-Object-Cache
  */
@@ -365,12 +365,22 @@ class WP_Object_Cache {
 			'retry_interval' => 0,
 		);
 
+		if ( defined( 'WP_REDIS_IGNORED_GROUPS' ) && is_array( WP_REDIS_IGNORED_GROUPS ) ) {
+			$this->ignored_groups = array_unique( array_merge( $this->ignored_groups, array_map( array( $this, 'sanitize_key' ), WP_REDIS_IGNORED_GROUPS ) ) );
+		}
+
 		$client = 'predis';
 		if ( class_exists( 'Redis' ) && 0 !== strcasecmp( 'predis', $client ) ) {
 			$client = defined( 'HHVM_VERSION' ) ? 'hhvm' : 'pecl';
 		}
 
+		$scheme = '/' === substr( $parameters['host'], 0, 1 ) ? 'unix' : 'tcp';
+
 		try {
+			if ( 'unix' === $scheme && ( 'hhvm' === $client || 'pecl' === $client ) ) {
+				$parameters['port'] = null;
+			}
+
 			if ( 'hhvm' === $client ) {
 				$this->redis_client = sprintf( 'HHVM Extension (v%s)', constant( 'HHVM_VERSION' ) );
 
@@ -396,6 +406,10 @@ class WP_Object_Cache {
 
 			if ( ( 'hhvm' === $client || 'pecl' === $client ) && defined( 'WPHB_REDIS_PASSWORD' ) ) {
 				$this->redis->auth( WPHB_REDIS_PASSWORD );
+			}
+
+			if ( ( 'hhvm' === $client || 'pecl' === $client ) && defined( 'WPHB_REDIS_DB_ID' ) ) {
+				$this->redis->select( WPHB_REDIS_DB_ID );
 			}
 
 			if ( 'predis' === $client ) {
@@ -426,6 +440,13 @@ class WP_Object_Cache {
 
 				$options = array();
 
+				if ( 'unix' === $scheme ) {
+					$parameters['scheme'] = $scheme;
+					$parameters['path']   = $parameters['host'];
+					unset( $parameters['host'] );
+					unset( $parameters['port'] );
+				}
+
 				if ( $parameters['read_timeout'] ) {
 					$parameters['read_write_timeout'] = $parameters['read_timeout'];
 				}
@@ -434,7 +455,12 @@ class WP_Object_Cache {
 					$options['parameters']['password'] = WPHB_REDIS_PASSWORD;
 				}
 
+				if ( defined( 'WPHB_REDIS_DB_ID' ) ) {
+					$options['parameters']['database'] = WPHB_REDIS_DB_ID;
+				}
+
 				$this->redis = new Predis\Client( $parameters, $options );
+
 				$this->redis->connect();
 
 				$this->redis_client .= sprintf( ' (v%s)', Predis\Client::VERSION );
@@ -539,9 +565,7 @@ class WP_Object_Cache {
 	 * @return bool Returns TRUE on success or FALSE on failure.
 	 */
 	protected function add_or_replace( $add, $key, $value, $group = 'default', $expiration = 0 ) {
-		$addition_suspended = function_exists( 'wp_suspend_cache_addition' )
-			? wp_suspend_cache_addition()
-			: false;
+		$addition_suspended = function_exists( 'wp_suspend_cache_addition' ) && wp_suspend_cache_addition();
 
 		if ( $add && $addition_suspended ) {
 			return false;
@@ -551,7 +575,7 @@ class WP_Object_Cache {
 		$derived_key = $this->build_key( $key, $group );
 
 		// Save if group not excluded and redis is up.
-		if ( ! in_array( $group, $this->ignored_groups ) && $this->redis_status() ) {
+		if ( ! in_array( $group, $this->ignored_groups, true ) && $this->redis_status() ) {
 			$exists = $this->redis->exists( $derived_key );
 
 			if ( $add == $exists ) {
@@ -596,7 +620,7 @@ class WP_Object_Cache {
 			$result = true;
 		}
 
-		if ( $this->redis_status() && ! in_array( $group, $this->ignored_groups ) ) {
+		if ( $this->redis_status() && ! in_array( $group, $this->ignored_groups, true ) ) {
 			$result = $this->parse_redis_response( $this->redis->del( $derived_key ) );
 		}
 
@@ -604,8 +628,7 @@ class WP_Object_Cache {
 	}
 
 	/**
-	 * Invalidate all items in the cache. If `WP_REDIS_SELECTIVE_FLUSH` is `true`,
-	 * only keys prefixed with the `WP_CACHE_KEY_SALT` are flushed.
+	 * Invalidate all items in the cache.
 	 *
 	 * @return bool Returns TRUE on success or FALSE on failure.
 	 */
@@ -614,7 +637,12 @@ class WP_Object_Cache {
 		$this->cache = array();
 
 		if ( $this->redis_status() ) {
-			$results[] = $this->parse_redis_response( $this->redis->flushdb() );
+			if ( defined( 'WP_REDIS_SALT' ) && WP_REDIS_SALT ) {
+				$lua_script = $this->get_flush_closure( trim( WP_REDIS_SALT ) );
+				$results[]  = $this->parse_redis_response( $lua_script() );
+			} else {
+				$results[] = $this->parse_redis_response( $this->redis->flushdb() );
+			}
 		}
 
 		if ( empty( $results ) ) {
@@ -633,7 +661,7 @@ class WP_Object_Cache {
 	/**
 	 * Returns a closure to flush selectively.
 	 *
-	 * @param  string $salt The salt to be used to differentiate.
+	 * @param  string $salt  The salt to be used to differentiate.
 	 * @return callable      Generated callable executing the lua script.
 	 */
 	protected function get_flush_closure( $salt ) {
@@ -647,7 +675,7 @@ class WP_Object_Cache {
 	/**
 	 * Returns a closure ready to be called to flush selectively ignoring unflushable groups.
 	 *
-	 * @param  string $salt The salt to be used to differentiate.
+	 * @param  string $salt  The salt to be used to differentiate.
 	 * @return callable      Generated callable executing the lua script.
 	 */
 	protected function lua_flush_closure( $salt ) {
@@ -691,7 +719,7 @@ LUA;
 		return function () use ( $salt ) {
 			$salt_length = strlen( $salt );
 
-			$unflushable = array_map(
+			$non_flushable = array_map(
 				function ( $group ) {
 					return ":{$group}:";
 				},
@@ -726,8 +754,8 @@ LUA;
 			}
 
 			$args = ( $this->redis instanceof Predis\Client )
-				? array_merge( array( $script, count( $unflushable ) ), $unflushable )
-				: array( $script, $unflushable, count( $unflushable ) );
+				? array_merge( array( $script, count( $non_flushable ) ), $non_flushable )
+				: array( $script, $non_flushable, count( $non_flushable ) );
 
 			return call_user_func_array( array( $this->redis, 'eval' ), $args );
 		};
@@ -753,7 +781,7 @@ LUA;
 			$found = true;
 			$this->cache_hits++;
 			return is_object( $this->cache[ $derived_key ] ) ? clone $this->cache[ $derived_key ] : $this->cache[ $derived_key ];
-		} elseif ( in_array( $group, $this->ignored_groups ) || ! $this->redis_status() ) {
+		} elseif ( in_array( $group, $this->ignored_groups, true ) || ! $this->redis_status() ) {
 			$found = false;
 			$this->cache_misses++;
 			return false;
@@ -784,7 +812,7 @@ LUA;
 	 * Usage: array( 'group0' => array( 'key0', 'key1', 'key2', ), 'group1' => array( 'key0' ) ).
 	 *
 	 * @param  array $groups Array of groups and keys to retrieve.
-	 * @return bool|mixed Array of cached values, keys in the format $group:$key. Non-existent keys null.
+	 * @return array|false Array of cached values, keys in the format $group:$key. Non-existent keys null.
 	 */
 	public function get_multi( $groups ) {
 		if ( empty( $groups ) || ! is_array( $groups ) ) {
@@ -795,7 +823,7 @@ LUA;
 		$cache = array();
 
 		foreach ( $groups as $group => $keys ) {
-			if ( in_array( $group, $this->ignored_groups ) || ! $this->redis_status() ) {
+			if ( in_array( $group, $this->ignored_groups, true ) || ! $this->redis_status() ) {
 				foreach ( $keys as $key ) {
 					$cache[ $this->build_key( $key, $group ) ] = $this->get( $key, $group );
 				}
@@ -852,7 +880,7 @@ LUA;
 		$derived_key = $this->build_key( $key, $group );
 
 		// Save if group not excluded from redis and redis is up.
-		if ( ! in_array( $group, $this->ignored_groups ) && $this->redis_status() ) {
+		if ( ! in_array( $group, $this->ignored_groups, true ) && $this->redis_status() ) {
 			$expiration = $this->validate_expiration( $expiration );
 
 			if ( $expiration ) {
@@ -883,7 +911,7 @@ LUA;
 		$offset      = (int) $offset;
 
 		// If group is a non-Redis group, save to internal cache, not Redis.
-		if ( in_array( $group, $this->ignored_groups ) || ! $this->redis_status() ) {
+		if ( in_array( $group, $this->ignored_groups, true ) || ! $this->redis_status() ) {
 			$value  = $this->get_from_internal_cache( $derived_key, $group );
 			$value += $offset;
 			$this->add_to_internal_cache( $derived_key, $value );
@@ -911,7 +939,7 @@ LUA;
 		$offset      = (int) $offset;
 
 		// If group is a non-Redis group, save to internal cache, not Redis.
-		if ( in_array( $group, $this->ignored_groups ) || ! $this->redis_status() ) {
+		if ( in_array( $group, $this->ignored_groups, true ) || ! $this->redis_status() ) {
 			$value  = $this->get_from_internal_cache( $derived_key, $group );
 			$value -= $offset;
 			$this->add_to_internal_cache( $derived_key, $value );
@@ -939,8 +967,8 @@ LUA;
 			$group = 'default';
 		}
 
-		$salt   = defined( 'WP_CACHE_KEY_SALT' ) ? trim( WP_CACHE_KEY_SALT ) : '';
-		$prefix = in_array( $group, $this->global_groups ) ? $this->global_prefix : $this->blog_prefix;
+		$salt   = defined( 'WP_REDIS_SALT' ) ? trim( WP_REDIS_SALT ) . ':' : '';
+		$prefix = in_array( $group, $this->global_groups, true ) ? $this->global_prefix : $this->blog_prefix;
 
 		$key   = str_replace( ':', '-', $key );
 		$group = str_replace( ':', '-', $group );
@@ -971,7 +999,8 @@ LUA;
 	 * Convert Redis responses into something meaningful
 	 *
 	 * @param  mixed $response Redis response.
-	 * @return mixed
+	 *
+	 * @return bool|int|string
 	 */
 	protected function parse_redis_response( $response ) {
 		if ( is_bool( $response ) ) {
@@ -1115,7 +1144,7 @@ LUA;
 	 *
 	 * @return mixed Unserialized data can be any type.
 	 */
-	function maybe_unserialize( $original ) {
+	public function maybe_unserialize( $original ) {
 		if ( $this->is_serialized( $original ) ) { // Don't attempt to unserialize data that wasn't serialized going in.
 			return @unserialize( $original );
 		}
@@ -1206,8 +1235,22 @@ LUA;
 	 * @return int
 	 */
 	protected function validate_expiration( $expiration ) {
-		$expiration = is_int( $expiration ) || ctype_digit( $expiration ) ? (int) $expiration : 0;
-		return $expiration;
+		return is_int( $expiration ) || ctype_digit( $expiration ) ? (int) $expiration : 0;
+	}
+
+	/**
+	 * Sanitizes a string key.
+	 *
+	 * Keys are used as internal identifiers. Lowercase alphanumeric characters,
+	 * dashes, and underscores are allowed.
+	 *
+	 * @param string $key  String key.
+	 *
+	 * @return string Sanitized key.
+	 */
+	public function sanitize_key( $key ) {
+		$key = strtolower( $key );
+		return preg_replace( '/[^a-z0-9_\-]/', '', $key );
 	}
 
 }
